feat: cart 기능 추가, order 기능 추가
- cart 아이템 추가
- cart 아이템 수량 감소
- cart 아이템 삭제
- cart 목록 조회
- cart 아이템 하나 전체 조회

- 단건 주문
- 카트 다건 주문
- 주문 목록 조회
- 주문 상세보기

+++ src/main/java/com/app/preorder/entity/product/Product.java
@@ -8,6 +8,7 @@ import lombok.*;
 import org.antlr.v4.runtime.misc.NotNull;
 import org.springframework.ws.soap.addressing.server.annotation.Address;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -27,7 +28,7 @@ public class Product {
     private String productName;
 
     @NotNull
-    private Long productPrice;
+    private BigDecimal productPrice;
 
     @NotNull
     private String description;
